Increase buffer size to support maximum message length
All data from UART is copied directly into this stack variable. The eCTF message format reserves 2 bytes for the packet length, meaning the maximum message size is 0xffff. This variable is increased to that +1 to hold all the data. Note that this means half of the SRAM is dedicated to this one stack variable, but it compiles and works fine. An alternate solution is to set this buffer size to a specific value and check the length in host_messaging.c's read_bytes() function.