Check size of DECODE data
In the reference design, the frame consists of a channel, timestamp, and actual data. The reference assumes that the length of the data is large enough for a channel and timestamp. However, if a malicious DECODE packet is sent that is SHORTER than the size of a channel and timestamp, `frame_size` will underflow and become an extremely large number, causing extra bytes to be printed to the console from memory. 

This check is added to ensure that the length of the data provided in the DECODE packet is AT LEAST the size of the channel and timestamp, otherwise it throws an error and returns early. I also have it error if `pkt_len` is the same size as the channel and timestamp, because that means a frame of 0 bytes was sent to us, and that shouldn't be a thing.