Check data length of subscription update packet
The reference design casts the `uart_buffer` variable into a pointer to a subscription update packet, assuming that there's enough data in there to fill the entire `subscription_update_packet_t` struct. However, this is never checked. A check is now added at the beginning of `update_subscription()` to ensure that the length of the payload is the SAME size as a `subscription_update_packet_t` struct, otherwise it errors out. This ensures that uninitialized, leftover data in memory is being used to fill in this subscription. 

While this code is likely to change in our implementations, I want to add this check in anyways so developers know to check it.